chore: Refactored mod names
