+++ mobile/native/src/lib.rs
@@ -1,7 +1,7 @@
 mod api;
-pub mod api_calculations;
-pub mod api_lndlc;
-pub mod api_model;
 mod bridge_generated;
+pub mod calculations;
+pub mod ln_dlc;
 pub mod logger;
+pub mod model;
 pub mod trade;
